Ui: Improve abstraction in PersonListPanel
We separate out the keyboard event handler and click event handler to PersonListPanel.
