+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -4,15 +4,11 @@ import java.util.Stack;
 import java.util.logging.Logger;
 
 import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextInputControl;
-import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
-import javafx.scene.input.MouseButton;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 import seedu.address.commons.core.GuiSettings;
@@ -124,32 +120,7 @@ public class MainWindow extends UiPart<Stage> {
      * Fills up all the placeholders of this window.
      */
     void fillInnerParts() {
-        personListPanel = new PersonListPanel(logic.getFilteredPersonList());
-
-        EventHandler keyEventHandler = new EventHandler<KeyEvent>() {
-            @Override
-            public void handle(KeyEvent event) {
-                if (event.getCode().equals(KeyCode.ENTER)) {
-                    Person selectedPerson = personListPanel.getSelectedPerson();
-                    switchMainPanel(new DetailPanel(selectedPerson), true);
-                }
-            }
-        };
-        personListPanel.setKeyEventHandler(keyEventHandler);
-
-        EventHandler clickEventHandler = new EventHandler<MouseEvent>() {
-            @Override
-            public void handle(MouseEvent event) {
-                // Double Click
-                if (event.getButton().equals(MouseButton.PRIMARY)) {
-                    if (event.getClickCount() == 2) {
-                        Person selectedPerson = personListPanel.getSelectedPerson();
-                        switchMainPanel(new DetailPanel(selectedPerson), true);
-                    }
-                }
-            }
-        };
-        personListPanel.setClickEventHandler(clickEventHandler);
+        personListPanel = new PersonListPanel(logic.getFilteredPersonList(), this::switchPersonDetailPanel);
 
         currentMainPanel = personListPanel;
         mainPanelPlaceholder.getChildren().add(personListPanel.getRoot());
@@ -176,6 +147,13 @@ public class MainWindow extends UiPart<Stage> {
         }
     }
 
+    /**
+     * Switch between different main panels
+     *
+     * @param panel to be switched
+     * @param recordHistory flag to indicate whether this action need
+     *                      to be stored
+     */
     private void switchMainPanel(MainPanel panel, boolean recordHistory) {
         if (currentMainPanel == panel) {
             return;
@@ -257,5 +235,10 @@ public class MainWindow extends UiPart<Stage> {
             resultDisplay.setFeedbackToUser(e.getMessage());
             throw e;
         }
+    }
+
+    private void switchPersonDetailPanel(Person person) {
+        Person selectedPerson = personListPanel.getSelectedPerson();
+        switchMainPanel(new DetailPanel(selectedPerson), true);
     }
 }
